requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-text-splitters
langchain-huggingface
langchain_chroma
langchain-core
sentence-transformers
openai
redis
aioredis
chromadb
python-dotenv
pydantic
httpx
websockets